+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Tests
+
+- Dispatcher service tests for the `dispatch_workflow.py` module.
+
 ## [0.53.0] - 2022-04-04
 
 ### Added